requirements.txt
Django==5.2
google_api_python_client==2.169.0
google_auth_oauthlib==1.2.2
ninja==1.11.1.4
protobuf==6.31.1
python-dotenv==1.1.0
Requests==2.32.3
spotipy==2.25.1
